backend-python/requirements.txt
requests>=2.28.0
openai>=0.27.0
python-dotenv>=1.0.0
